proj/requirements.txt
fastapi==0.118.3
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
pymongo==4.10.1
neo4j==5.28.2
llama-index==0.14.4
langchain==0.3.27
langchain-community==0.3.31
langchain-ollama==0.3.10
ollama==0.6.0
langfuse==3.8.1
PyMuPDF==1.26.4
streamlit==1.50.0
pymongo==4.10.1
huggingface-hub==0.35.3
transformers==4.57.0
pydantic==2.12.0
google-generativeai==0.8.5

